feat: stream logs to debug file and migrate commands to node-pty
