+++ package.json
@@ -13,11 +13,12 @@
     "@aws-sdk/client-sts": "^3.588.0",
     "@aws-sdk/credential-providers": "^3.588.0",
     "@swc/helpers": "~0.5.2",
-    "chalk": "4.1.2",
+    "chalk": "^4.1.2",
     "command-exists": "^1.2.9",
     "ejs": "^3.1.10",
     "enquirer": "^2.4.1",
     "log-symbols": "^4.1.0",
+    "node-pty": "^1.0.0",
     "ora": "^5.4.1",
     "pino": "^9.1.0",
     "pino-pretty": "^11.1.0",
