+++ packages/cli/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "vm-x",
-  "version": "0.2.0",
+  "name": "vm-x-ai",
+  "version": "0.2.3",
   "description": "VM-X CLI",
   "license": "MIT",
   "bin": {
-    "vm-x": "./index.cjs.js"
+    "vm-x-ai": "./bin/index.cjs.js"
   },
   "repository": {
     "type": "git",
@@ -29,13 +29,15 @@
     "@aws-sdk/client-sts": "^3.588.0",
     "@aws-sdk/credential-providers": "^3.588.0",
     "@smithy/types": "^3.0.0",
-    "chalk": "4.1.2",
+    "chalk": "^4.1.2",
     "log-symbols": "^4.1.0",
     "ejs": "^3.1.10",
     "yargs": "^17.7.2",
     "pino": "^9.1.0",
+    "pino-pretty": "^11.1.0",
     "command-exists": "^1.2.9",
-    "enquirer": "^2.4.1"
+    "enquirer": "^2.4.1",
+    "node-pty": "^1.0.0"
   },
   "main": "./src/index.js",
   "typings": "./src/index.d.ts"
