+++ packages/cli/rollup.config.js
@@ -1,4 +1,11 @@
 module.exports = (config) => {
-  // config.output.preserveModules = true;
+  config.input = {
+    index: 'packages/cli/src/index.ts',
+    'bin/index': 'packages/cli/bin/index.ts',
+  };
+  for (const output of config.output) {
+    output.preserveModules = true;
+  }
+
   return config;
 };
